Develop

+++ app/Http/Controllers/MessengerController.php
@@ -7,16 +7,29 @@ use App\Http\Requests\MessageUpdateRequest;
 use App\Models\Message;
 use App\Models\User;
 use App\Services\MessengerControllerService;
+use Illuminate\Mail\Events\MessageSent;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Notification;
+use App\Notifications\SendedMessage;
 
 class MessengerController
 {
+    protected $user;
+
+    public function __construct() {
+        $this->user = Auth::user();
+    }
+
     /*
     *  Получить список комментариев между мельзователями
     */
-    public function getListOfUsers(User $sender, User $receiver, MessengerControllerService $service): array
+    public function getListOfUsers(User $receiver, MessengerControllerService $service)
     {
+        $sender = $this->user;
         $message = $service->getListOfUser($sender, $receiver);
-
+        Log::info(['Отправитель - ' => $sender->id, 'Получатель' => $receiver->id]);
         return [
             'count' => count($message),
             'items' => $message
@@ -26,12 +39,14 @@ class MessengerController
     /*
     *  Отправить сообщение пользователю
     */
-    public function send(User $sender, User $receiver, MessageSendRequest $request, MessengerControllerService $service): array
+    public function send(User $receiver, MessageSendRequest $request, MessengerControllerService $service): array
     {
+        $sender = $this->user;
+        Gate::authorize('create', Message::class);
 
         $message = $service->createFromRequest($sender, $receiver, $request->validated());
 
-        abort_unless($message, 404);
+        $receiver->notify(new SendedMessage($sender));
 
         return [
             'status' => 'success',
@@ -44,8 +59,9 @@ class MessengerController
     */
     public function update(Message $message, MessageUpdateRequest $request, MessengerControllerService $service): array
     {
+        Gate::authorize('update', $message);
         $message = $service->updateFromRequest($message, $request->validated());
-        abort_unless($message, 500);
+
         return [
             'status' => 'success',
             'message' => $message
@@ -57,8 +73,9 @@ class MessengerController
      */
     public function delete(Message $message, MessengerControllerService $service): array
     {
+        Gate::authorize('delete', $message);
         $service->delete($message);
-        abort_unless($message, 500);
+
         return [
             'status' => 'success',
         ];
